create http_server module
